style(feature/DomainLayer): Group files in folders
